Fixed thread ui/data separation issue. Moved only the angle modification to the gameThread, left the position modification in the UI thread. Also moved camera displacement to the UI thread. Added an update method to signify UI thread changes. Run method is for gameThread changes. Should probably change this to a more suitable pattern. Also changed sizes of bodies independent to sprite size for the creator. The BodyObject handles conversion from sprite size to data size.

+++ powerup-core/src/joozey/libs/powerup/game/GameData.java
@@ -86,9 +86,9 @@ public class GameData
 	}
 
 
-    public static float getSpeed()
+    public static double getSpeed()
     {
-        return 3f;
+        return 3.;
     }
 
     public static void setSpeed( float speed )
